backend/requirements.txt
# Backend Python dependencies for Proximity 2.0
# Last updated: 2025-10-31 (Final Security Patch)
# Security: All critical vulnerabilities patched to latest versions

# Django Core (5.1.x - latest with ALL security patches)
Django==5.1.13
djangorestframework==3.15.2
django-ninja==1.3.0
django-cors-headers==4.4.0
django-extensions==3.2.3
Werkzeug==3.1.3
pyOpenSSL==24.2.1

# Database
psycopg2-binary>=2.9.9
dj-database-url==2.1.0

# Authentication & Security
dj-rest-auth==6.0.0
django-allauth==0.62.0
djangorestframework-simplejwt==5.5.1
PyJWT==2.9.0
bcrypt==4.1.3
cryptography>=44.0.1
python-dotenv==1.0.1

# Async & Background Tasks
celery==5.4.0
redis==5.2.0
channels==4.2.0
channels-redis==4.2.0
autobahn==24.10.1
daphne==4.1.2

# Proxmox Integration
proxmoxer==2.1.1
requests>=2.32.4

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-dateutil>=2.8.2
Pillow>=10.3.0
PyYAML>=6.0.1
paramiko>=3.5.0

# Monitoring & Logging
sentry-sdk[django]>=2.0.0

# Testing & Development
pytest>=7.4.4
pytest-django>=4.7.0
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
faker>=22.1.0

# Code Quality
black>=24.3.0
isort>=5.13.2
flake8>=7.0.0
bandit>=1.7.5
pip-audit>=2.6.0
